--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -2,10 +2,15 @@ module github.com/ar2r/go-otus/hw10_program_optimization
 
 go 1.22
 
-require github.com/stretchr/testify v1.7.0
+require (
+	github.com/json-iterator/go v1.1.12
+	github.com/stretchr/testify v1.7.0
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
+	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )