--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,11 +1,17 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
+	"bufio"
 	"fmt"
 	"io"
-	"regexp"
 	"strings"
+
+	jsoniter "github.com/json-iterator/go" //nolint:depguard
+)
+
+var (
+	json             = jsoniter.ConfigCompatibleWithStandardLibrary
+	ErrInvalidDomain = fmt.Errorf("invalid domain")
 )
 
 type User struct {
@@ -21,46 +27,39 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := io.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
 	result := make(DomainStat)
 
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
+	domain = strings.Trim(domain, " ")
+	if domain == "" {
+		return result, ErrInvalidDomain
+	}
+
+	scanner := bufio.NewScanner(r)
+	for scanner.Scan() {
+		userEmail, err := unmarshalUserEmail(scanner.Bytes())
 		if err != nil {
-			return nil, err
+			return result, err
 		}
 
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+		foundDomain := ExtractMatchedDomain(userEmail, domain)
+		if foundDomain != "" {
+			result[foundDomain]++
 		}
 	}
 	return result, nil
+}
+
+func unmarshalUserEmail(data []byte) (string, error) {
+	var user User
+	if err := json.Unmarshal(data, &user); err != nil {
+		return "", fmt.Errorf("unmarshal user error: %w", err)
+	}
+	return user.Email, nil
+}
+
+func ExtractMatchedDomain(email, domain string) string {
+	if strings.Contains(email, domain) {
+		return strings.SplitN(strings.ToLower(email), "@", 2)[1]
+	}
+	return ""
 }