--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -16,7 +16,9 @@ func TestGetDomainStat(t *testing.T) {
 {"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
 {"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
 {"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}
-{"Id":6,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@почта.рф","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
+{"Id":6,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@почта.рф","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}
+{"Id":7,"Name":"Janice Rose","Username":"KeithHart","Email":"com-not-email","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}
+{"Id":8,"Name":"Janice Rose","Username":"KeithHart","Email":"com@mail.ru","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
 
 	t.Run("find 'com'", func(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "com")