## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

```
# Запустил на коде из master ветки
go test -bench=BenchmarkGetDomainStat -benchmem  -count=10 > old.txt
# Запустил на оптимизированном коде
go test -bench=BenchmarkGetDomainStat -benchmem  -count=10 > new.txt

# Результат оптимизации
benchstat old.txt new.txt                                           
goos: darwin
goarch: arm64
pkg: github.com/ar2r/go-otus/hw10_program_optimization
cpu: Apple M2 Pro
                 │     old.txt     │               new.txt               │
                 │     sec/op      │   sec/op     vs base                │
GetDomainStat-12   68729.163µ ± 0%   5.490µ ± 5%  -99.99% (p=0.000 n=10)

                 │      old.txt      │                new.txt                │
                 │       B/op        │     B/op      vs base                 │
GetDomainStat-12   132859.625Ki ± 0%   6.589Ki ± 0%  -100.00% (p=0.000 n=10)

                 │    old.txt     │              new.txt               │
                 │   allocs/op    │ allocs/op   vs base                │
GetDomainStat-12   1700101.0 ± 0%   106.0 ± 0%  -99.99% (p=0.000 n=10)
```

### Критерии оценки
- [x] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [x] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов